Fix broken "Awesomeness" org. structure
The Markdown syntax was invalid, using 2 spaces instead of the required 4 to indent sub-bullets.  This caused the Github formatter to become confused and think that the "Awesomeness" requirements were not supposed to be nested.  These *are* supposed to be nested though, so fix this by fixing the spaces.

Fixes Issue #2