Add existing public Navy instruction as base for PR demos
Adding the OPNAVINST 6110.1J (reformatted hastily in Markdown) to serve as a base for examples on how to update the published instruction based on later changes (e.g. NAVADMINs).  Enclosure (2) and after are missing for now, and the paragraph numbering format is different (though the numbers themselves should be the same).